# `ultralytics 8.3.77` faster YOLOv8-Segment ONNX Runtime example