# **YOLOv8 ONNXRUNTIME Python Segmentation Optimization**  

## **Overview**  
This optimization focused on improving the overall inference time of YOLOv8 segmentation, mainly by optimizing post-processing. Performance was measured over **1001 iterations**, with the **first execution dropped** to eliminate initialization overhead.  

## **Results**  

## **GPU Performance**
| Stage            | **Before (ms)** | **After (ms)** | **Improvement** |
|-----------------|--------------|-------------|--------------|
| **Preprocess**  | 1.00  | 0.85  | **15% Faster** |
| **Postprocess** | 15.87 | 1.31  | **91.7% Faster** |
| **Total**       | 48.42 | 32.73 | **32.4% Faster** |

## **CPU Performance**
| Stage            | **Before (ms)** | **After (ms)** | **Improvement** |
|-----------------|--------------|-------------|--------------|
| **Preprocess**  | 1.00  | 1.00  | **No Change** |
| **Postprocess** | 13.91 | 5.72  | **58.9% Faster** |
| **Total**       | 311.46 | 253.62 | **18.6% Faster** |

## **Summary**
- **Post-processing was the main bottleneck** and saw the biggest improvements.  
- **GPU total inference time reduced by 32.4%** and **CPU by 18.6%**.  
- **First execution was dropped** for accurate benchmarking.   


## **Benchmarking Methodology**  

To accurately measure the performance improvements, the **`perf_counter`** from Python's `time` module was utilized. This high-resolution timer provides precise time measurements, making it ideal for performance testing. 

### **Implementation Details**
- **Benchmark Code**: The benchmarking code was structured to encapsulate the preprocessing, post-processing, and total inference stages. 
- **Iterations**: The performance was evaluated over **1001 iterations** to ensure statistical significance, with the first execution excluded to mitigate the effects of initialization overhead.
- **Profiling Stages**: Each stage of the YOLOv8 segmentation pipeline was individually timed to identify bottlenecks and assess the impact of optimizations.

By employing `perf_counter`, the results reflect a more accurate assessment of the optimization efforts, allowing for clear comparisons between the before and after states of the inference process.


#### OLD VERSION BENCHMARK

![old_1](https://github.com/user-attachments/assets/2502c257-6a6e-49f2-b188-8275e3ec09ea)
![old_2](https://github.com/user-attachments/assets/eda8c4d4-3839-41ad-9e4d-e83813c43dea)

#### NEW VERSION BENCHMARK

![new_1](https://github.com/user-attachments/assets/ea638b95-625d-43e8-bd56-e68d4ee2f2c0)
![new_2](https://github.com/user-attachments/assets/66ca8e58-c740-4b32-9822-6a7ccf124fb2)


I have read the CLA Document and I sign the CLA

<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
--->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Updated the YOLOv8 segmentation ONNXRuntime Python example script for cleaner structure and added new functionalities. 🛠️

### 📊 Key Changes
- **Enhanced Model Initialization**:
  - Added a `confidence threshold` parameter for improved flexibility during segment filtering.
  - Defined new attributes like `device` to specify CPU/GPU inference and dataset class names.
- **Code Restructuring**:
  - Simplified and modularized code for pre-processing, post-processing, and mask handling.
  - Removed unused parameters (e.g., IoU threshold) from function calls in favor of streamlined logic.
- **Post-Processing Upgrade**:
  - Integrated PyTorch-based non-max suppression and mask handling for better performance and clarity.
  - Added automatic scale and mask cropping methods for result refinement.
- **Output Visualization**:
  - Replaced custom visualization logic with a simpler `cv2.imshow()` for real-time display of results.

### 🎯 Purpose & Impact
- **Improved Functionality**:
  - Provides users greater control over model confidence thresholds.
  - Introduces a more reusable and structured framework for ONNX-based segmentation.
- **Enhanced Usability**:
  - Simplifies both inference and result visualization, making it easier for users to test models.
- **Broader Compatibility**:
  - Supports GPU/CPU flexibility and handles different ONNX model configurations (e.g., FP16 or FP32).
- **Cleaner Codebase**:
  - More modular and maintainable for future development or integration into larger pipelines. 🚀