@AdnanEkici 2025-02-19
# **YOLOv8 Segmentation Optimization Report**  

## **Overview**  
This optimization focused on improving the overall inference time of YOLOv8 segmentation, mainly by optimizing post-processing. Performance was measured over **1001 iterations**, with the **first execution dropped** to eliminate initialization overhead.  

## **Results**  

## **GPU Performance**
| Stage            | **Before (ms)** | **After (ms)** | **Improvement** |
|-----------------|--------------|-------------|--------------|
| **Preprocess**  | 1.00  | 0.85  | **15% Faster** |
| **Postprocess** | 15.87 | 1.31  | **91.7% Faster** |
| **Total**       | 48.42 | 32.73 | **32.4% Faster** |

## **CPU Performance**
| Stage            | **Before (ms)** | **After (ms)** | **Improvement** |
|-----------------|--------------|-------------|--------------|
| **Preprocess**  | 1.00  | 1.00  | **No Change** |
| **Postprocess** | 13.91 | 5.72  | **58.9% Faster** |
| **Total**       | 311.46 | 253.62 | **18.6% Faster** |

## **Summary**
- **Post-processing was the main bottleneck** and saw the biggest improvements.  
- **GPU total inference time reduced by 32.4%** and **CPU by 18.6%**.  
- **First execution was dropped** for accurate benchmarking.   


## **Benchmarking Methodology**  

To accurately measure the performance improvements, the **`perf_counter`** from Python's `time` module was utilized. This high-resolution timer provides precise time measurements, making it ideal for performance testing. 

### **Implementation Details**
- **Benchmark Code**: The benchmarking code was structured to encapsulate the preprocessing, post-processing, and total inference stages. 
- **Iterations**: The performance was evaluated over **1001 iterations** to ensure statistical significance, with the first execution excluded to mitigate the effects of initialization overhead.
- **Profiling Stages**: Each stage of the YOLOv8 segmentation pipeline was individually timed to identify bottlenecks and assess the impact of optimizations.

By employing `perf_counter`, the results reflect a more accurate assessment of the optimization efforts, allowing for clear comparisons between the before and after states of the inference process.


#### OLD VERSION BENCHMARK

![old_1](https://github.com/user-attachments/assets/2502c257-6a6e-49f2-b188-8275e3ec09ea)
![old_2](https://github.com/user-attachments/assets/eda8c4d4-3839-41ad-9e4d-e83813c43dea)

#### NEW VERSION BENCHMARK

![new_1](https://github.com/user-attachments/assets/ea638b95-625d-43e8-bd56-e68d4ee2f2c0)
![new_2](https://github.com/user-attachments/assets/66ca8e58-c740-4b32-9822-6a7ccf124fb2)