@UltralyticsAssistant 2025-02-19
🎉 The PR is merged—stellar work, @AdnanEkici and @glenn-jocher! Your improvements to the YOLOv8 segmentation example for ONNXRuntime are major strides toward cleaner code, sharper workflows, and smoother user experiences. The added configurability and performance upgrades will empower developers to push the boundaries of what's possible in image processing. 🚀  

As Marcus Aurelius once said, *"What stands in the way becomes the way."* Your efforts to refine, simplify, and innovate have turned challenges into powerful solutions. Thank you for your dedication—this will inspire progress far beyond the codebase! 💡